Make SDK root path configurable in the Makefiles
Compile the firmware images without making assumptions about the
location of the SDK relative to the Makefile.

For example, to compile te right side from the folder
redox-w-keyboard-basic:

$> SDK_ROOT=${HOME}/redox/nRF5_SDK_11.0.0_89a8197 make keyboard_side=right -C custom/armgcc/

Note that by default and if the user doesn't set the SDK_ROOT, the
build system makes the assumption that the redox-w-fireware is inside
the SDK_ROOT to avoid breaking backwards compatibility.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@gmail.com>