Add delay between setting a column pin and reading out the row
Currently, the way we do matrix scans in the two halfs can give rise
to race conditions. In the loop we set a column pin and in the next
cycle we read the the row. As a result we might be reading the input
before the singals stabilize. This change adds a nop that allows a
cycle between the two operations.

In addition, this patch remove a workaround that made the local
variable volatile in an attempt to fix this.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@gmail.com>