Power off instead of disabling the rtc (#19)
Previously, after 0.5sec of inactivity we would disable the rtc and go
into a low power mode the next time we received a keypress which would
turn on the rtc and resume normal operation. This change simplifies
the code and instead switches system off. A key press will power it on
again and resume operation.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@gmail.com>