Receiver fixes and improvements (#22)
* Remove inactivity detection on the receiver side

We define the keyboard half as inactive when there is no keypress and
the matrix is empty. Instead of guessing at the receiver side that a
half is inactive after not sending an update for some time, make the
half send an empty matrix. When the keyboard is inactive this will be
the last update before powering off.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@gmail.com>

* Make receiver interrupt safe

If a packet arrives while the receiver is sending the matrix through
the uart, it updates the matrix and depending on timing, we might end
up sending to qmk a corrupt matrix. This patch changes the flow to
make this process safe.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@gmail.com>